Be/6-Money 클래스 도입

+++ backend/src/main/java/com/back/back9/domain/wallet/dto/CoinAmountResponse.java
@@ -20,7 +20,7 @@ public record CoinAmountResponse(
                 coin.getSymbol(),
                 coin.getKoreanName() != null ? coin.getKoreanName() : coin.getSymbol(),
                 coinAmount.getQuantity(),
-                coinAmount.getTotalAmount()
+                coinAmount.getTotalAmount().toBigDecimal()
         );
     }
 }
