+++ backend/src/main/java/com/back/back9/domain/wallet/dto/WalletBalanceResponse.java
@@ -27,7 +27,7 @@ public record  WalletBalanceResponse (
                 wallet.getId(),
                 wallet.getUser().getId(),
                 wallet.getAddress(),
-                wallet.getBalance()
+                wallet.getBalance().toBigDecimal()
         );
     }
